# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "STX",
        "contact": {},
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/executions/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "execution"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "公共执行编号",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/execution.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/executions/{id}/cancel": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "execution"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "公共执行编号",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "幂等键",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "确认取消影响",
                        "name": "X-STX-Confirm",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/execution.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/executions/{id}/wait": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "execution"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "公共执行编号",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "服务端单次等待秒数，范围为 1 到 30",
                        "name": "timeout_seconds",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/execution.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/oauth/callback": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oauth"
                ],
                "parameters": [
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/oauth.CallbackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/oauth.CallbackResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/oauth/login": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oauth"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth provider (github, google)",
                        "name": "provider",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/oauth.GetLoginURLResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "execution.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error_code": {
                    "type": "string"
                },
                "error_msg": {
                    "type": "string"
                }
            }
        },
...
        "oauth.CallbackRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "oauth.CallbackResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error_msg": {
                    "type": "string"
                }
            }
        },
        "oauth.GetLoginURLResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "error_msg": {
                    "type": "string"
                }
            }
        },
...
    }
}